back to deepspeed half model

+++ train_scrc.py
@@ -209,7 +209,7 @@ for i in range(2):
 
 model = utils.initialize_model(args.classifier,
                                num_classes=n_classes,
-                               chn_dim=len(scrc_in)).half().to(device)
+                               chn_dim=len(scrc_in)).to(device)
 optimizer = optim.Adam(model.parameters(), lr=args.lr)
 criterion = torch.nn.CrossEntropyLoss()
 
@@ -237,7 +237,7 @@ for epoch in range(args.begin_epoch, args.nepochs):
         bat_id = np.random.rand(x.shape[0]).argsort()
         x = x[bat_id, ]
         y = y[bat_id, ]
-        x = x.half().to(device)
+        x = x.to(device)
         y = y.to(device)
 
         logits = model(x)
@@ -258,7 +258,7 @@ for epoch in range(args.begin_epoch, args.nepochs):
     model.eval()
     tot, cor = 0, 0
     for _, (x, y) in enumerate(tst_loader[1]):
-        x = x.half().to(device)
+        x = x.to(device)
         y = y.to(device)
         lgts = model(x)
         _, pred = lgts.max(1)
